refactor: serializers_rpc module
