+++ ietf/api/views_rpc.py
@@ -1,11 +1,7 @@
 # Copyright The IETF Trust 2023, All Rights Reserved
 
-import datetime
-from typing import Literal, Optional
-
 from django.db.models.functions import Coalesce
-from drf_spectacular.types import OpenApiTypes
-from drf_spectacular.utils import extend_schema_view, extend_schema, OpenApiParameter, extend_schema_field
+from drf_spectacular.utils import extend_schema_view, extend_schema, OpenApiParameter
 from rest_framework import serializers, viewsets, mixins
 from rest_framework.decorators import action
 from rest_framework.views import APIView
@@ -17,25 +13,21 @@ from django.http import (
 )
 
 from ietf.doc.factories import WgDraftFactory  # DO NOT MERGE INTO MAIN
-from ietf.doc.models import Document, DocHistory, DocumentAuthor
+from ietf.doc.models import Document, DocHistory
 from ietf.person.factories import PersonFactory  # DO NOT MERGE INTO MAIN
 from ietf.person.models import Person
-
-
-class PersonBatchSerializer(serializers.Serializer):
-    person_ids = serializers.ListField(child=serializers.IntegerField())
-
-
-class PersonSerializer(serializers.ModelSerializer):
-    """Serializer for a Person in a response"""
-    plain_name = serializers.SerializerMethodField()
-
-    class Meta:
-        model = Person
-        fields = ["id", "plain_name"]
-
-    def get_plain_name(self, person) -> str:
-        return person.plain_name()
+from .serializers_rpc import (
+    PersonBatchSerializer,
+    PersonSerializer,
+    FullDraftSerializer,
+    DraftSerializer,
+    SubmittedToQueueSerializer,
+    OriginalStreamSerializer,
+    DemoPersonCreateSerializer,
+    DemoPersonSerializer,
+    DemoDraftCreateSerializer,
+    DemoDraftSerializer,
+)
 
 
 @extend_schema_view(
@@ -50,7 +42,7 @@ class PersonSerializer(serializers.ModelSerializer):
         description="returns a dict of person pks to person names",
         request=PersonBatchSerializer,
         responses=PersonSerializer(many=True),
-    )
+    ),
 )
 class PersonViewSet(mixins.RetrieveModelMixin, viewsets.GenericViewSet):
     queryset = Person.objects.all()
@@ -62,7 +54,9 @@ class PersonViewSet(mixins.RetrieveModelMixin, viewsets.GenericViewSet):
     def batch(self, request):
         """Get a batch of rpc person names"""
         pks = PersonBatchSerializer(request.data).data["person_ids"]
-        return Response(self.get_serializer(Person.objects.filter(pk__in=pks), many=True).data)
+        return Response(
+            self.get_serializer(Person.objects.filter(pk__in=pks), many=True).data
+        )
 
 
 class SubjectPersonView(APIView):
@@ -86,95 +80,13 @@ class SubjectPersonView(APIView):
         try:
             user_id = int(subject_id)
         except ValueError:
-            raise serializers.ValidationError({"subject_id": "This field must be an integer value."})
+            raise serializers.ValidationError(
+                {"subject_id": "This field must be an integer value."}
+            )
         person = Person.objects.filter(user__pk=user_id).first()
         if person:
             return Response(PersonSerializer(person).data)
         raise Http404
-
-
-class DocumentAuthorSerializer(serializers.ModelSerializer):
-    """Serializer for a Person in a response"""
-    plain_name = serializers.SerializerMethodField()
-
-    class Meta:
-        model = DocumentAuthor
-        fields = ["person", "plain_name"]
-
-    def get_plain_name(self, document_author: DocumentAuthor) -> str:
-        return document_author.person.plain_name()
-
-
-class FullDraftSerializer(serializers.ModelSerializer):
-    source_format = serializers.SerializerMethodField()
-    authors = DocumentAuthorSerializer(many=True, source="documentauthor_set")
-    shepherd = serializers.SerializerMethodField()
-
-    class Meta:
-        model = Document
-        fields = [
-            "id",
-            "name",
-            "rev",
-            "stream",
-            "title",
-            "pages",
-            "source_format",
-            "authors",
-            "shepherd",
-            "intended_std_level",
-        ]
-
-    def get_source_format(self, doc: Document) -> Literal["unknown", "xml-v2", "xml-v3", "txt"]:
-        submission = doc.submission()
-        if submission is None:
-            return "unknown"
-        if ".xml" in submission.file_types:
-            if submission.xml_version == "3":
-                return "xml-v3"
-            else:
-                return "xml-v2"
-        elif ".txt" in submission.file_types:
-            return "txt"
-        return "unknown"
-
-    @extend_schema_field(OpenApiTypes.EMAIL)
-    def get_shepherd(self, doc: Document) -> str:
-        if doc.shepherd:
-            return doc.shepherd.formatted_ascii_email()
-        return ""
-
-
-class DraftSerializer(FullDraftSerializer):
-    class Meta:
-        model = Document
-        fields = [
-            "id",
-            "name",
-            "rev",
-            "stream",
-            "title",
-            "pages",
-            "source_format",
-            "authors",
-        ]
-
-
-class SubmittedToQueueSerializer(FullDraftSerializer):
-    submitted = serializers.SerializerMethodField()
-
-    class Meta:
-        model = Document
-        fields = [
-            "id",
-            "name",
-            "stream",
-            "submitted",
-        ]
-
-    def get_submitted(self, doc) -> Optional[datetime.datetime]:
-        event = doc.sent_to_rfc_editor_event()
-        return None if event is None else event.time
 
 
 @extend_schema_view(
@@ -188,7 +100,7 @@ class SubmittedToQueueSerializer(FullDraftSerializer):
         summary="List documents ready to enter the RFC Editor Queue",
         description="List documents ready to enter the RFC Editor Queue",
         responses=SubmittedToQueueSerializer(many=True),
-    )
+    ),
 )
 class DraftViewSet(mixins.RetrieveModelMixin, viewsets.GenericViewSet):
     queryset = Document.objects.filter(type_id="draft")
@@ -212,19 +124,13 @@ class DraftViewSet(mixins.RetrieveModelMixin, viewsets.GenericViewSet):
             states__slug__in=["rfc-edit"],
         )
         # TODO: Need a way to talk about editorial stream docs
-        docs = self.get_queryset().filter(type_id="draft").filter(
-            ietf_docs | irtf_iab_ise_docs
+        docs = (
+            self.get_queryset()
+            .filter(type_id="draft")
+            .filter(ietf_docs | irtf_iab_ise_docs)
         )
         serializer = self.get_serializer(docs, many=True)
         return Response(serializer.data)
-
-
-class OriginalStreamSerializer(serializers.ModelSerializer):
-    stream = serializers.CharField(read_only=True, source="orig_stream_id")
-
-    class Meta:
-        model = Document
-        fields = ["rfc_number", "stream"]
 
 
 @extend_schema_view(
@@ -265,39 +171,12 @@ class DraftsByNamesView(APIView):
         summary="Get a batch of drafts by draft names",
         description="returns a list of drafts with matching names",
         request=list[str],
-        responses=DraftSerializer(many=True)
+        responses=DraftSerializer(many=True),
     )
     def post(self, request):
         names = request.data
         docs = Document.objects.filter(type_id="draft", name__in=names)
         return Response(DraftSerializer(docs, many=True).data)
-
-
-class DemoPersonCreateSerializer(serializers.Serializer):
-    name = serializers.CharField(max_length=255)
-
-
-class DemoPersonSerializer(serializers.ModelSerializer):
-    person_pk = serializers.IntegerField(source="pk")
-
-    class Meta:
-        model = Person
-        fields = ["user_id", "person_pk"]
-
-
-class DemoDraftCreateSerializer(serializers.Serializer):
-    name = serializers.CharField(max_length=255, required=True)
-    stream_id = serializers.CharField(default="ietf")
-    rev = serializers.CharField(default=None)
-    states = serializers.DictField(child=serializers.CharField(), default=None)
-
-
-class DemoDraftSerializer(serializers.ModelSerializer):
-    doc_id = serializers.IntegerField(source="pk")
-
-    class Meta:
-        model = Document
-        fields = ["doc_id", "name"]
 
 
 @extend_schema_view(
@@ -312,13 +191,14 @@ class DemoDraftSerializer(serializers.ModelSerializer):
         operation_id="create_demo_draft",
         summary="Build a datatracker WG draft for RPC demo purposes",
         description="returns a datatracker document id for a draft created with the provided name and states. "
-                    "The arguments, if present, are passed directly to the WgDraftFactory",
+        "The arguments, if present, are passed directly to the WgDraftFactory",
         request=DemoDraftCreateSerializer,
         responses=DemoDraftSerializer,
     ),
 )
 class DemoViewSet(viewsets.ViewSet):
     """SHOULD NOT MAKE IT INTO PRODUCTION"""
+
     api_key_endpoint = "ietf.api.views_rpc"
 
     @action(detail=False, methods=["post"])
@@ -347,7 +227,9 @@ class DemoViewSet(viewsets.ViewSet):
             doc = WgDraftFactory(
                 **kwargs
             )  # Yes, things may be a little strange if the stream isn't IETF, but until we need something different...
-            event_type = "iesg_approved" if stream_id == "ietf" else "requested_publication"
+            event_type = (
+                "iesg_approved" if stream_id == "ietf" else "requested_publication"
+            )
             if not doc.docevent_set.filter(
                 type=event_type
             ).exists():  # Not using get_or_create here on purpose - these are wobbly facades we're creating
